# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TryNex Lifestyle API — Bangladesh custom merchandise e-commerce platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /promo-codes/validate:
    post:
      operationId: validatePromoCode
      tags: [promo-codes]
      summary: Validate a promo or referral code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                orderTotal: { type: number }
                customerEmail: { type: string, format: email }
              required: [code, orderTotal]
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  discount: { type: number }
                  code: { type: string }
                  message: { type: string }
                  isReferral: { type: boolean }
                required: [valid]
  /promo-codes/{id}:
    patch:
      operationId: updatePromoCode
      tags: [promo-codes]
      summary: Partially update a promo code (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                active: { type: boolean }
                discountType: { type: string, enum: [percentage, fixed] }
                discountValue: { type: number }
                minOrderAmount: { type: number }
                maxUses: { type: integer }
                expiresAt: { type: string, format: date-time, nullable: true }
      responses:
        "200":
          description: Updated promo code
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      operationId: deletePromoCode
      tags: [promo-codes]
      summary: Delete a promo code (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
  /referrals:
    get:
      operationId: listReferrals
      tags: [referrals]
      summary: List referral codes (admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Referral list
          content:
            application/json:
              schema:
                type: object
                properties:
                  referrals:
                    type: array
                    items: { type: object }
                required: [referrals]
  /referrals/{id}:
    patch:
      operationId: updateReferral
      tags: [referrals]
      summary: Toggle referral active status (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                active: { type: boolean }
              required: [active]
      responses:
        "200":
          description: Updated referral
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      operationId: deleteReferral
      tags: [referrals]
      summary: Delete a referral (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...